requirements.txt
pandas>=2.0,<3.0
numpy>=1.24,<2.0
tensorflow-cpu>=2.15,<3.0
scikit-learn>=1.3,<2.0
streamlit>=1.37,<2.0
streamlit-folium>=0.18,<1.0
folium>=0.15,<1.0
plotly>=5.18,<6.0
